requirements.txt
openai>=1.0.0
requests
beautifulsoup4
python-dotenv
